Initial example to allow an external S3 bucket;
Providing this variable in a test module will cause the upload of local
files to the bootstrap bucket to not happen. There are other issues with
this approach currently. We should probably not create the bootstrap
bucket to begin with if the user is providing its own. Also we should
document that when `external_bootstrap_bucket` is set, it means that
files must be in the external S3 bucket, e.g. licenses and proxy certs.