Update the tests/public-install outputs;
Now that we're running the K6 tests, we need to construct the URLs for
various endpoints. I'm purposefully putting these in the output for the
tests/public-install module since it doesn't need to be in the root
module. Then, we construct these endpoints in the output and leverage
them in the GitHub actions for `handler-test.yml` rather than
hard-coding paths in the Action's YAML.