Add k6 Smoke Tests (#129)
* Update docs for Checkout of tests/

* Update the output from root to be the tfe_url

* Update the tests/public-install outputs;

Now that we're running the K6 tests, we need to construct the URLs for
various endpoints. I'm purposefully putting these in the output for the
tests/public-install module since it doesn't need to be in the root
module. Then, we construct these endpoints in the output and leverage
them in the GitHub actions for `handler-test.yml` rather than
hard-coding paths in the Action's YAML.

* Refactor the working directory stanza

* Add in the necessary steps for K6 to work

* Add in the output for K6 smoke tests

* Retrieve the create-admin output

Co-authored-by: Aaron Lane <10655063+aaron-lane@users.noreply.github.com>

* Fix grammar

Co-authored-by: Aaron Lane <10655063+aaron-lane@users.noreply.github.com>

* Update YAML anchors to job environment variables;

The #128 PR added YAML anchors, but GitHub Actions doesn't support this.

```
The workflow is not valid. .github/workflows/handler-test.yml (Line: 4, Col: 1):
Unexpected value 'tfe_parameters' .github/workflows/handler-test.yml: Anchors
are not currently supported. Remove the anchor 'PUBLIC_WORK_DIR'
```

Co-authored-by: Aaron Lane <10655063+aaron-lane@users.noreply.github.com>