[TF-15186] add active-active as valid operational mode input (#334)
This module manages an "active-active" state using the var.node_count. This is different from the other modes, which are only based on the variable operational_mode.

This PR adds active-active as a valid value for var.operational_mode and the node_count won't affect the active-active status. This change is non-breaking, so consumers using active-active by providing a node_count greater than 2 won't be affected.